Adding option to compile phpdbg sapi
Because phpdbg is a sapi, it is a separate php binary to compile, not an
extension.

The formula for phpdbg is a source-only formula that simply copies the source
code for the sapi so that we can leverage normal Formula behaviors like
checksum-ing, downloading releases or from head, etc.

The new option for abstract-php allows us to pass a --with-phpdbg option to
trigger the necessary additional dependencies and add the necessary compilation
steps. This option also causes us to remove dtrace and libedit from the compile
flags per krakjoe/phpdbg#38.

Tested on my mbp with Mavericks and works fine except for avoidable segfault
per https://bugs.php.net/bug.php?id=66575